feat:PUT e DELETE OK

+++ src/main/java/com/vitoriaferreira/repositories/DeviceRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.vitoriaferreira.entities.Device;
 
-public interface DeviceRepository extends JpaRepository<Device, Long> {
+public interface DeviceRepository extends JpaRepository<Device, Integer> {
 
 }
